app/requirements.txt
alembic==1.13.1
anthropic==0.31.2
celery==5.4.0
fastapi==0.111.1
graphviz==0.20.3
instructor==1.4.0
ipython==8.31.0
pandas==2.2.3
psycopg2-binary==2.9.9
pydantic==2.10.4
pydantic-settings==2.7.0
python-frontmatter==1.1.0
redis==5.0.3
timescale-vector==0.0.7
docling
tiktoken
requests
ipykernel
python-dotenv
openai
httpx==0.27.0
pytest==8.3.3
